metadata can only be attached to instructions
See: https://llvm.org/docs/LangRef.html#metadata

Attempting to set metadata on other LLVMValueRef will cause undefined
behaviour in llvm, see here for example:

https://github.com/llvm/llvm-project/blob/release/8.x/llvm/lib/IR/Core.cpp#L851
