+++ src/values/ptr_value.rs
@@ -5,7 +5,7 @@ use std::ffi::CStr;
 
 use crate::support::LLVMString;
 use crate::types::{AsTypeRef, IntType, PointerType};
-use crate::values::{AsValueRef, InstructionValue, IntValue, Value, MetadataValue};
+use crate::values::{AsValueRef, InstructionValue, IntValue, Value};
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
 pub struct PointerValue {
@@ -66,18 +66,6 @@ impl PointerValue {
 
     pub fn as_instruction(&self) -> Option<InstructionValue> {
         self.ptr_value.as_instruction()
-    }
-
-    pub fn has_metadata(&self) -> bool {
-        self.ptr_value.has_metadata()
-    }
-
-    pub fn get_metadata(&self, kind_id: u32) -> Option<MetadataValue> {
-        self.ptr_value.get_metadata(kind_id)
-    }
-
-    pub fn set_metadata(&self, metadata: MetadataValue, kind_id: u32) {
-        self.ptr_value.set_metadata(metadata, kind_id)
     }
 
     // REVIEW: Should this be on array value too?
