+++ src/values/struct_value.rs
@@ -5,7 +5,7 @@ use std::ffi::CStr;
 use crate::support::LLVMString;
 use crate::types::StructType;
 use crate::values::traits::AsValueRef;
-use crate::values::{InstructionValue, Value, MetadataValue};
+use crate::values::{InstructionValue, Value};
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
 pub struct StructValue {
@@ -51,18 +51,6 @@ impl StructValue {
 
     pub fn as_instruction(&self) -> Option<InstructionValue> {
         self.struct_value.as_instruction()
-    }
-
-    pub fn has_metadata(&self) -> bool {
-        self.struct_value.has_metadata()
-    }
-
-    pub fn get_metadata(&self, kind_id: u32) -> Option<MetadataValue> {
-        self.struct_value.get_metadata(kind_id)
-    }
-
-    pub fn set_metadata(&self, metadata: MetadataValue, kind_id: u32) {
-        self.struct_value.set_metadata(metadata, kind_id)
     }
 
     pub fn replace_all_uses_with(&self, other: StructValue) {
