+++ src/values/vec_value.rs
@@ -6,7 +6,7 @@ use std::ffi::CStr;
 use crate::support::LLVMString;
 use crate::types::{VectorType};
 use crate::values::traits::AsValueRef;
-use crate::values::{BasicValueEnum, BasicValue, InstructionValue, Value, IntValue, MetadataValue};
+use crate::values::{BasicValueEnum, BasicValue, InstructionValue, Value, IntValue};
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
 pub struct VectorValue {
@@ -99,18 +99,6 @@ impl VectorValue {
         };
 
         BasicValueEnum::new(value)
-    }
-
-    pub fn has_metadata(&self) -> bool {
-        self.vec_value.has_metadata()
-    }
-
-    pub fn get_metadata(&self, kind_id: u32) -> Option<MetadataValue> {
-        self.vec_value.get_metadata(kind_id)
-    }
-
-    pub fn set_metadata(&self, metadata: MetadataValue, kind_id: u32) {
-        self.vec_value.set_metadata(metadata, kind_id)
     }
 
     pub fn replace_all_uses_with(&self, other: VectorValue) {
